feat: seo optimization
